+++ src/Dialog/Dialog.tsx
@@ -5,10 +5,10 @@
  */
 
 import React, { ReactNode } from 'react';
+import { Spinner } from 'react-bootstrap';
 import Modal from 'react-bootstrap/Modal';
 
 import Button, { ButtonVariants } from '../Button/Button';
-import Spinner from './Spinner';
 
 import './dialog.scss';
 
@@ -16,25 +16,32 @@ type CoreProps = {
     isVisible: boolean;
     onHide?: () => void;
     className?: string;
-    size?: 'sm' | 'lg' | 'xl';
+    size?: 'sm' | 'm' | 'lg' | 'xl';
     children: ReactNode | string;
 };
 
 type DialogProps = CoreProps & {
     closeOnUnfocus?: boolean;
+    closeOnEsc?: boolean;
 };
 
 export const Dialog = ({
     isVisible,
     closeOnUnfocus = false,
+    closeOnEsc = false,
     onHide = () => {},
     className = '',
-    size = 'lg',
+    size = 'm',
     children,
 }: DialogProps) => (
     <Modal
+        onEscapeKeyDown={() => {
+            if (closeOnEsc && onHide) {
+                onHide();
+            }
+        }}
         show={isVisible}
-        size={size}
+        size={size === 'm' ? undefined : size}
         backdrop={closeOnUnfocus ? true : 'static'}
         onHide={() => {
             if (closeOnUnfocus && onHide) {
@@ -51,13 +58,16 @@ export const Dialog = ({
 Dialog.Header = ({
     title,
     headerIcon,
+    showSpinner,
 }: {
     title: string;
     headerIcon?: string;
+    showSpinner?: boolean;
 }) => (
     <Modal.Header closeButton={false}>
         <p>
             <b>{title}</b>
+            {showSpinner && <Spinner size="sm" animation="border" />}
         </p>
         {headerIcon && <span className={`mdi mdi-${headerIcon}`} />}
     </Modal.Header>
@@ -67,18 +77,17 @@ Dialog.Body = ({ children }: { children: ReactNode | string }) => (
     <Modal.Body>{children}</Modal.Body>
 );
 
-Dialog.Footer = ({
-    showSpinner = false,
-    children,
-}: {
-    showSpinner?: boolean;
-    children: ReactNode;
-}) => (
-    <Modal.Footer>
-        {showSpinner && <Spinner />}
-        {children}
-    </Modal.Footer>
+Dialog.Footer = ({ children }: { children: ReactNode }) => (
+    <Modal.Footer>{children}</Modal.Footer>
 );
+
+export interface DialogButtonProps {
+    onClick: () => void;
+    variant?: ButtonVariants;
+    className?: string;
+    disabled?: boolean;
+    children: ReactNode | string;
+}
 
 export const DialogButton = ({
     variant = 'secondary',
@@ -86,13 +95,7 @@ export const DialogButton = ({
     className = '',
     disabled = false,
     children,
-}: {
-    onClick: React.MouseEventHandler<HTMLButtonElement>;
-    variant?: ButtonVariants;
-    className?: string;
-    disabled?: boolean;
-    children: ReactNode | string;
-}) => (
+}: DialogButtonProps) => (
     <Button
         large
         variant={variant}
@@ -104,33 +107,73 @@ export const DialogButton = ({
     </Button>
 );
 
+interface GenericDialogProps extends CoreProps {
+    title: string;
+    footer: React.ReactNode;
+    headerIcon?: string;
+    showSpinner?: boolean;
+    closeOnUnfocus?: boolean;
+    closeOnEsc?: boolean;
+}
+
+export const GenericDialog = ({
+    isVisible,
+    onHide,
+    title,
+    headerIcon,
+    children,
+    className,
+    footer,
+    showSpinner = false,
+    closeOnUnfocus,
+    closeOnEsc,
+    size,
+}: GenericDialogProps) => (
+    <Dialog
+        onHide={onHide}
+        closeOnUnfocus={closeOnUnfocus}
+        closeOnEsc={closeOnEsc}
+        isVisible={isVisible}
+        className={className}
+        size={size}
+    >
+        <Dialog.Header
+            title={title}
+            headerIcon={headerIcon}
+            showSpinner={showSpinner}
+        />
+        <Dialog.Body>{children}</Dialog.Body>
+        <Dialog.Footer>{footer}</Dialog.Footer>
+    </Dialog>
+);
+
 interface InfoProps extends CoreProps {
     title?: string;
     headerIcon?: string;
+    onHide: () => void;
 }
 
 export const InfoDialog = ({
     isVisible,
     title = 'Info',
-    headerIcon,
+    headerIcon = 'info',
     children,
-    onHide = () => {},
-    size = 'lg',
+    onHide,
+    size,
     className,
 }: InfoProps) => (
-    <Dialog
-        isVisible={isVisible}
-        closeOnUnfocus
+    <GenericDialog
+        closeOnEsc
         onHide={onHide}
+        isVisible={isVisible}
+        headerIcon={headerIcon}
+        title={title}
         className={className}
         size={size}
+        footer={<DialogButton onClick={onHide}>Close</DialogButton>}
     >
-        <Dialog.Header title={title} headerIcon={headerIcon} />
-        <Dialog.Body>{children}</Dialog.Body>
-        <Dialog.Footer>
-            <DialogButton onClick={() => onHide()}>Close</DialogButton>
-        </Dialog.Footer>
-    </Dialog>
+        {children}
+    </GenericDialog>
 );
 
 export const ErrorDialog = (props: Omit<InfoProps, 'headerIcon'>) =>
@@ -162,22 +205,31 @@ export const ConfirmationDialog = ({
     cancelLabel = 'Cancel',
     onCancel,
     optionalLabel,
-    onOptional,
-    size = 'lg',
+    onOptional = () => {},
+    size,
 }: ConfirmationDialogProps) => (
-    <Dialog isVisible={isVisible} className={className} size={size}>
-        <Dialog.Header title={title} headerIcon={headerIcon} />
-        <Dialog.Body>{children}</Dialog.Body>
-        <Dialog.Footer>
-            {onOptional && optionalLabel && (
-                <DialogButton onClick={onOptional}>
-                    {optionalLabel}
+    <GenericDialog
+        onHide={onCancel}
+        closeOnEsc
+        isVisible={isVisible}
+        headerIcon={headerIcon}
+        title={title}
+        className={className}
+        size={size}
+        footer={
+            <>
+                <DialogButton variant="primary" onClick={onConfirm}>
+                    {confirmLabel}
                 </DialogButton>
-            )}
-            <DialogButton onClick={onConfirm} variant="primary">
-                {confirmLabel}
-            </DialogButton>
-            <DialogButton onClick={onCancel}>{cancelLabel}</DialogButton>
-        </Dialog.Footer>
-    </Dialog>
+                <DialogButton onClick={onCancel}>{cancelLabel}</DialogButton>
+                {optionalLabel && (
+                    <DialogButton onClick={onOptional}>
+                        {optionalLabel}
+                    </DialogButton>
+                )}
+            </>
+        }
+    >
+        {children}
+    </GenericDialog>
 );