+++ src/index.ts
@@ -14,12 +14,13 @@ export { default as Logo } from './Logo/Logo';
 export { default as DeviceSelector } from './Device/DeviceSelector/DeviceSelector';
 export {
     Dialog,
+    GenericDialog,
     ConfirmationDialog,
     InfoDialog,
     ErrorDialog,
     DialogButton,
 } from './Dialog/Dialog';
-export { default as Spinner } from './Dialog/Spinner';
+export type { DialogButtonProps } from './Dialog/Dialog';
 export { default as Slider } from './Slider/Slider';
 export { Toggle } from './Toggle/Toggle';
 export { default as RootErrorDialog } from './ErrorDialog/ErrorDialog';